+++ server/Simulation/Components/PickableComponent.cs
@@ -1,0 +1,6 @@
+using server.ECS;
+
+namespace server.Simulation.Components;
+
+[Component]
+public readonly struct PickableTagComponent() { }